refactor: Hotel-1 move DI objects to their own modules
