+++ data/src/main/java/uz/otamurod/data/interactor/HotelInteractor.kt
@@ -1,11 +1,11 @@
 package uz.otamurod.data.interactor
 
 import kotlinx.coroutines.flow.Flow
-import uz.otamurod.domain.api.model.Booking
-import uz.otamurod.domain.api.model.Hotel
-import uz.otamurod.domain.api.model.Rooms
 import uz.otamurod.domain.api.repository.HotelRepositoryApi
 import uz.otamurod.domain.interactor.HotelInteractorApi
+import uz.otamurod.domain.model.Booking
+import uz.otamurod.domain.model.Hotel
+import uz.otamurod.domain.model.Rooms
 import uz.otamurod.domain.util.DataState
 import javax.inject.Inject
