+++ data/src/main/java/uz/otamurod/data/network/mapper/AboutHotelResponseMapper.kt
@@ -1,7 +1,7 @@
 package uz.otamurod.data.network.mapper
 
 import uz.otamurod.data.network.api.entities.AboutHotelResponse as AboutHotelDto
-import uz.otamurod.domain.api.model.AboutHotel as AboutHotelBo
+import uz.otamurod.domain.model.AboutHotel as AboutHotelBo
 
 object AboutHotelResponseMapper {
     class AboutHotel internal constructor(private val dto: AboutHotelDto){
