+++ data/src/main/java/uz/otamurod/data/network/mapper/BookingResponseMapper.kt
@@ -1,7 +1,7 @@
 package uz.otamurod.data.network.mapper
 
 import uz.otamurod.data.network.api.entities.BookingResponse as BookingDto
-import uz.otamurod.domain.api.model.Booking as BookingBo
+import uz.otamurod.domain.model.Booking as BookingBo
 
 object BookingResponseMapper {
     class Booking internal constructor(private val dto: BookingDto){
