+++ data/src/main/java/uz/otamurod/data/network/mapper/RoomResponseMapper.kt
@@ -1,7 +1,7 @@
 package uz.otamurod.data.network.mapper
 
 import uz.otamurod.data.network.api.entities.RoomResponse as RoomDto
-import uz.otamurod.domain.api.model.Room as RoomBo
+import uz.otamurod.domain.model.Room as RoomBo
 
 object RoomResponseMapper {
     class Room internal constructor(private val dto: RoomDto) {
