+++ data/src/main/java/uz/otamurod/data/network/mapper/RoomsResponseMapper.kt
@@ -1,7 +1,7 @@
 package uz.otamurod.data.network.mapper
 
 import uz.otamurod.data.network.api.entities.RoomsResponse as RoomsDto
-import uz.otamurod.domain.api.model.Rooms as RoomsBo
+import uz.otamurod.domain.model.Rooms as RoomsBo
 
 object RoomsResponseMapper {
     class Rooms internal constructor(private val dto: RoomsDto) {
