+++ data/src/main/java/uz/otamurod/data/repository/HotelRepository.kt
@@ -1,14 +1,12 @@
 package uz.otamurod.data.repository
 
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOn
 import uz.otamurod.data.remote.datasource.HotelRemoteDataSource
-import uz.otamurod.domain.api.model.Booking
-import uz.otamurod.domain.api.model.Hotel
-import uz.otamurod.domain.api.model.Rooms
 import uz.otamurod.domain.api.repository.HotelRepositoryApi
+import uz.otamurod.domain.model.Booking
+import uz.otamurod.domain.model.Hotel
+import uz.otamurod.domain.model.Rooms
 import uz.otamurod.domain.util.DataState
 import javax.inject.Inject
 
@@ -21,7 +19,7 @@ class HotelRepository @Inject constructor(
             emit(DataState.loading())
             val result = hotelRemoteDataSource.fetchHotel()
             emit(result)
-        }.flowOn(Dispatchers.IO)
+        }
     }
 
     override suspend fun fetchHotelRooms(): Flow<DataState<Rooms>> {
@@ -29,7 +27,7 @@ class HotelRepository @Inject constructor(
             emit(DataState.loading())
             val result = hotelRemoteDataSource.fetchHotelRooms()
             emit(result)
-        }.flowOn(Dispatchers.IO)
+        }
     }
 
     override suspend fun fetchBookingInfo(): Flow<DataState<Booking>> {
@@ -37,6 +35,6 @@ class HotelRepository @Inject constructor(
             emit(DataState.loading())
             val result = hotelRemoteDataSource.fetchBookingInfo()
             emit(result)
-        }.flowOn(Dispatchers.IO)
+        }
     }
 }
