+++ domain/src/main/java/uz/otamurod/domain/api/repository/HotelRepositoryApi.kt
@@ -1,9 +1,9 @@
 package uz.otamurod.domain.api.repository
 
 import kotlinx.coroutines.flow.Flow
-import uz.otamurod.domain.api.model.Booking
-import uz.otamurod.domain.api.model.Hotel
-import uz.otamurod.domain.api.model.Rooms
+import uz.otamurod.domain.model.Booking
+import uz.otamurod.domain.model.Hotel
+import uz.otamurod.domain.model.Rooms
 import uz.otamurod.domain.util.DataState
 
 interface HotelRepositoryApi {
