+++ domain/src/main/java/uz/otamurod/domain/interactor/HotelInteractorApi.kt
@@ -1,9 +1,9 @@
 package uz.otamurod.domain.interactor
 
 import kotlinx.coroutines.flow.Flow
-import uz.otamurod.domain.api.model.Booking
-import uz.otamurod.domain.api.model.Hotel
-import uz.otamurod.domain.api.model.Rooms
+import uz.otamurod.domain.model.Booking
+import uz.otamurod.domain.model.Hotel
+import uz.otamurod.domain.model.Rooms
 import uz.otamurod.domain.util.DataState
 
 interface HotelInteractorApi {
