+++ domain/src/main/java/uz/otamurod/domain/model/Rooms.kt
@@ -1,4 +1,4 @@
-package uz.otamurod.domain.api.model
+package uz.otamurod.domain.model
 
 data class Rooms(
     val rooms: List<Room>
