+++ presentation/src/main/java/uz/otamurod/presentation/ui/hotel/HotelViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
-import uz.otamurod.domain.api.model.Hotel
 import uz.otamurod.domain.interactor.HotelInteractorApi
+import uz.otamurod.domain.model.Hotel
 import uz.otamurod.domain.util.DataState
 import javax.inject.Inject
