+++ presentation/src/main/java/uz/otamurod/presentation/ui/number/NumberFragment.kt
@@ -14,8 +14,8 @@ import com.makeramen.roundedimageview.RoundedTransformationBuilder
 import com.squareup.picasso.Picasso
 import com.synnapps.carouselview.ImageListener
 import dagger.hilt.android.AndroidEntryPoint
-import uz.otamurod.domain.api.model.Room
-import uz.otamurod.domain.api.model.Rooms
+import uz.otamurod.domain.model.Room
+import uz.otamurod.domain.model.Rooms
 import uz.otamurod.domain.util.DataState
 import uz.otamurod.presentation.R
 import uz.otamurod.presentation.databinding.FragmentNumberBinding
